+++ packages/web/src/app/[lng]/(main)/grouping/create/funnels/main/MainStep.client.tsx
@@ -1,8 +1,15 @@
+import { format } from 'date-fns';
+
 import InputSection from './InputSection.client';
 import SettingSection from './SettingSection.client';
 import UploadSection from './UploadSection.client';
 import { useCreateGroupContext } from '../../components/CreateGroupContext';
 import CreateModal from '../../components/CreateModal.client';
+
+import type { CreateGroupContextValue } from '../../type';
+import type { TimeType } from '@/types';
+import type { SubmitHandler } from 'react-hook-form';
+
 import { useTranslation } from '@/app/i18n/client';
 import { Button, ButtonGroup } from '@/components/Button';
 import { Divider } from '@/components/Divider';
@@ -10,12 +17,7 @@ import { Spacing } from '@/components/Spacing';
 import { useDidMount } from '@/hooks/common/useDidMount';
 import useBrowser from '@/hooks/useBrowser';
 import { useModal, useToast } from '@/hooks/useModal';
-import sendMessageToReactNative from '@/utils/sendMessageToReactNative';
-import { format } from 'date-fns';
-
-import type { CreateGroupContextValue } from '../../type';
-import type { TimeType } from '@/types';
-import type { SubmitHandler } from 'react-hook-form';
+import { sendMessageToReactNative } from '@/utils';
 
 function validateDate(date: Date, time: TimeType, browser: string) {
   if (!date || !time.fromHour || !time.fromMin || !time.fromAmPm) {
