+++ packages/web/src/app/[lng]/(main)/profile/setting/langauge/components/Language.tsx
@@ -9,8 +9,7 @@ import { CircleCheckbox } from '@/components/Checkbox';
 import { Flex } from '@/components/Layout';
 import { Spacing } from '@/components/Spacing';
 import useAppRouter from '@/hooks/useAppRouter';
-import { getLocalCookie, setLocalCookie } from '@/utils/cookieController';
-import { afterDay60 } from '@/utils/date';
+import { afterDay60, getLocalCookie, setLocalCookie } from '@/utils';
 
 export default function Language() {
   const { t, i18n } = useTranslation('common');
