+++ packages/web/src/app/[lng]/(sub)/login/Login.tsx
@@ -1,10 +1,11 @@
 'use client';
 
+import { useEffect } from 'react';
+
 import { LoginResponse, useLoginMutation } from '@/apis/auth';
 import { Loading } from '@/components/Loading';
 import useAppRouter from '@/hooks/useAppRouter';
-import { setTokenAtCookie } from '@/utils/auth/tokenController';
-import { useEffect } from 'react';
+import { setTokenAtCookie } from '@/utils';
 
 export default function Login() {
   const { mutate: mutateLogin } = useLoginMutation();
