+++ packages/web/src/app/[lng]/page.tsx
@@ -4,9 +4,7 @@ import { cookieName } from '../i18n/settings';
 
 import { useDidMount } from '@/hooks/common/useDidMount';
 import useAppRouter from '@/hooks/useAppRouter';
-import { hasToken } from '@/utils/auth/tokenController';
-import { getLocalCookie, setLocalCookie } from '@/utils/cookieController';
-import { afterDay60 } from '@/utils/date';
+import { afterDay60, getLocalCookie, hasToken, setLocalCookie } from '@/utils';
 
 export default function Home() {
   const { i18n } = useTranslation('common');
