+++ packages/web/src/components/Analytics/GoogleAnalytics.client.tsx
@@ -1,10 +1,11 @@
 'use client';
 
-import { GA_TRACKING_ID } from '@/constants';
-import { pageview } from '@/utils/gtag';
 import { usePathname, useSearchParams } from 'next/navigation';
 import Script from 'next/script';
 import { useEffect } from 'react';
+
+import { GA_TRACKING_ID } from '@/constants';
+import { pageview } from '@/utils';
 
 export default function Analytics() {
   const pathname = usePathname();
