+++ packages/web/src/hooks/useCheckUpdate.ts
@@ -1,5 +1,4 @@
-import { getLocalCookie, setLocalCookie } from '@/utils/cookieController';
-import { getMobileDivce } from '@/utils/getMobileDevice';
+import { getLocalCookie, getMobileDivce, setLocalCookie } from '@/utils';
 
 export function useCheckUpdate(version: number) {
   const cookies = getLocalCookie('didUpdated_' + version);
